+++ src/main/java/com/kitisplode/golemfirststonemod/entity/client/model/EntityModelShieldFirstBrick.java
@@ -1,4 +1,4 @@
-package com.kitisplode.golemfirststonemod.entity.client;
+package com.kitisplode.golemfirststonemod.entity.client.model;
 
 import com.kitisplode.golemfirststonemod.GolemFirstStoneMod;
 import com.kitisplode.golemfirststonemod.entity.entity.effect.EntityEffectShieldFirstBrick;