+++ src/main/java/com/kitisplode/golemfirststonemod/entity/entity/golem/EntityGolemFirstDiorite.java
@@ -1,0 +1,217 @@
+package com.kitisplode.golemfirststonemod.entity.entity.golem;
+
+import com.kitisplode.golemfirststonemod.entity.ModEntities;
+import com.kitisplode.golemfirststonemod.entity.entity.IEntityWithDelayedMeleeAttack;
+import com.kitisplode.golemfirststonemod.entity.entity.golem.pawn.EntityPawnFirstDiorite;
+import com.kitisplode.golemfirststonemod.entity.goal.goal.MultiStageAttackGoalRanged;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.network.syncher.EntityDataAccessor;
+import net.minecraft.network.syncher.EntityDataSerializers;
+import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.util.Mth;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.AreaEffectCloud;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.ai.goal.*;
+import net.minecraft.world.entity.ai.goal.target.DefendVillageTargetGoal;
+import net.minecraft.world.entity.ai.goal.target.HurtByTargetGoal;
+import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
+import net.minecraft.world.entity.ai.goal.target.ResetUniversalAngerTargetGoal;
+import net.minecraft.world.entity.animal.IronGolem;
+import net.minecraft.world.entity.monster.Creeper;
+import net.minecraft.world.entity.monster.Enemy;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.Vec3;
+import software.bernie.geckolib.animatable.GeoEntity;
+import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
+import software.bernie.geckolib.core.animatable.instance.SingletonAnimatableInstanceCache;
+import software.bernie.geckolib.core.animation.AnimatableManager;
+import software.bernie.geckolib.core.animation.Animation;
+import software.bernie.geckolib.core.animation.AnimationController;
+import software.bernie.geckolib.core.animation.RawAnimation;
+
+
+public class EntityGolemFirstDiorite extends IronGolem implements GeoEntity, IEntityWithDelayedMeleeAttack
+{
+    private static final EntityDataAccessor<Integer> ATTACK_STATE = SynchedEntityData.defineId(EntityGolemFirstDiorite.class, EntityDataSerializers.INT);
+    private AnimatableInstanceCache cache = new SingletonAnimatableInstanceCache(this);
+    private final float attackRange = 32.0f;
+    private final int pawnsToSpawn = 4;
+
+    public EntityGolemFirstDiorite(EntityType<? extends IronGolem> pEntityType, Level pLevel)
+    {
+        super(pEntityType, pLevel);
+    }
+
+    public static AttributeSupplier setAttributes()
+    {
+        return Mob.createMobAttributes()
+                .add(Attributes.MAX_HEALTH, 1000.0D)
+                .add(Attributes.MOVEMENT_SPEED, 0.25f)
+                .add(Attributes.ATTACK_DAMAGE, 30.0f)
+                .add(Attributes.KNOCKBACK_RESISTANCE, 1.0f)
+                .add(Attributes.FOLLOW_RANGE, 48)
+                .build();
+    }
+
+    @Override
+    protected void defineSynchedData()
+    {
+        super.defineSynchedData();
+        this.entityData.define(ATTACK_STATE, 0);
+    }
+
+    public int getAttackState()
+    {
+        return this.entityData.get(ATTACK_STATE);
+    }
+
+    public void setAttackState(int pInt)
+    {
+        this.entityData.set(ATTACK_STATE, pInt);
+    }
+
+    private float getAttackDamage() {
+        return (float)this.getAttributeValue(Attributes.ATTACK_DAMAGE);
+    }
+
+    @Override
+    public double getEyeY()
+    {
+        return getY() + 2.2d;
+    }
+
+    @Override
+    protected void registerGoals()
+    {
+        this.goalSelector.addGoal(1, new MultiStageAttackGoalRanged(this, 1.0, true, Mth.square(attackRange), new int[]{300, 240, 50}));
+        this.goalSelector.addGoal(2, new MoveTowardsTargetGoal(this, 0.8D, 32.0F));
+        this.goalSelector.addGoal(2, new MoveBackToVillageGoal(this, 0.8D, false));
+        this.goalSelector.addGoal(4, new GolemRandomStrollInVillageGoal(this, 0.8D));
+        this.goalSelector.addGoal(7, new LookAtPlayerGoal(this, Player.class, 6.0F));
+        this.goalSelector.addGoal(8, new RandomLookAroundGoal(this));
+        this.targetSelector.addGoal(1, new DefendVillageTargetGoal(this));
+        this.targetSelector.addGoal(2, new HurtByTargetGoal(this));
+        this.targetSelector.addGoal(3, new NearestAttackableTargetGoal<>(this, Mob.class, 5, false, false, (entity) -> entity instanceof Enemy && !(entity instanceof Creeper)));
+        this.targetSelector.addGoal(4, new ResetUniversalAngerTargetGoal<>(this, false));
+    }
+
+    @Override
+    public boolean isPushable()
+    {
+        return getAttackState() == 0;
+    }
+
+    @Override
+    public boolean tryAttack()
+    {
+        if (getAttackState() != 3) return false;
+
+        this.level().broadcastEntityEvent(this, (byte)4);
+        this.playSound(SoundEvents.BEACON_POWER_SELECT, 1.0F, 1.0F);
+        spawnPawns(pawnsToSpawn);
+        return true;
+    }
+
+    private void spawnPawns(int pawnCount)
+    {
+        for (int i = 0; i < pawnCount; i++)
+        {
+            double direction = this.random.nextInt(360) * Mth.DEG_TO_RAD;
+            double offset = this.random.nextInt(4) + 2;
+            Vec3 spawnOffset = new Vec3(Math.sin(direction) * offset,
+                    0.0d,
+                    Math.cos(direction) * offset);
+            BlockState bs = level().getBlockState(new BlockPos((int)(getX() + spawnOffset.x()), (int)(getY() + spawnOffset.y()),(int)(getZ() + spawnOffset.z())));
+            int failCount = 0;
+            while (!bs.isAir() || bs.canOcclude())
+            {
+                spawnOffset = spawnOffset.add(0,1,0);
+                bs = level().getBlockState(new BlockPos((int)(getX() + spawnOffset.x()), (int)(getY() + spawnOffset.y()),(int)(getZ() + spawnOffset.z())));
+                failCount++;
+                if (failCount > 5) break;
+            }
+
+            EntityPawnFirstDiorite pawn = ModEntities.ENTITY_PAWN_FIRST_DIORITE.get().create(level());
+            if (pawn == null) continue;
+            pawn.setPlayerCreated(true);
+            pawn.moveTo(getX() + spawnOffset.x(), getY() + spawnOffset.y(), getZ() + spawnOffset.z(), 0.0f, 0.0F);
+            level().addFreshEntity(pawn);
+
+            AreaEffectCloud dust = new AreaEffectCloud(level(), getX() + spawnOffset.x(), getY() + spawnOffset.y(), getZ() + spawnOffset.z());
+            dust.setParticle(ParticleTypes.POOF);
+            dust.setRadius(1.0f);
+            dust.setDuration(1);
+            dust.setPos(dust.getX(),dust.getY(),dust.getZ());
+            level().addFreshEntity(dust);
+        }
+    }
+
+    @Override
+    protected InteractionResult mobInteract(Player pPlayer, InteractionHand pHand) {
+        ItemStack itemstack = pPlayer.getItemInHand(pHand);
+        if (!itemstack.is(Items.GOLD_INGOT)) {
+            return InteractionResult.PASS;
+        } else {
+            float f = this.getHealth();
+            this.heal(25.0F);
+            if (this.getHealth() == f) {
+                return InteractionResult.PASS;
+            } else {
+                float f1 = 1.0F + (this.random.nextFloat() - this.random.nextFloat()) * 0.2F;
+                this.playSound(SoundEvents.IRON_GOLEM_REPAIR, 1.0F, f1);
+                if (!pPlayer.getAbilities().instabuild) {
+                    itemstack.shrink(1);
+                }
+                return InteractionResult.sidedSuccess(this.level().isClientSide);
+            }
+        }
+    }
+
+    @Override
+    public void registerControllers(AnimatableManager.ControllerRegistrar controllerRegistrar)
+    {
+        controllerRegistrar.add(new AnimationController<>(this, "controller", 0, event ->
+        {
+            EntityGolemFirstDiorite pGolem = event.getAnimatable();
+            if (pGolem.getAttackState() > 0)
+            {
+                switch (pGolem.getAttackState())
+                {
+                    case 1:
+                        event.getController().setAnimationSpeed(0.5);
+                        return event.setAndContinue(RawAnimation.begin().then("animation.first_diorite.attack_windup", Animation.LoopType.HOLD_ON_LAST_FRAME));
+                    case 2:
+                        event.getController().setAnimationSpeed(1.00);
+                        return event.setAndContinue(RawAnimation.begin().then("animation.first_diorite.attack_charge", Animation.LoopType.LOOP));
+                    default:
+                        event.getController().setAnimationSpeed(1.00);
+                        return event.setAndContinue(RawAnimation.begin().then("animation.first_diorite.attack_end", Animation.LoopType.HOLD_ON_LAST_FRAME));
+                }
+            }
+            else
+            {
+                event.getController().setAnimationSpeed(1.00);
+                pGolem.setAttackState(0);
+                if (getDeltaMovement().horizontalDistanceSqr() > 0.001D || event.isMoving())
+                    return event.setAndContinue(RawAnimation.begin().thenLoop("animation.first_diorite.walk"));
+            }
+            return event.setAndContinue(RawAnimation.begin().thenLoop("animation.first_diorite.idle"));
+        }));
+    }
+
+    @Override public AnimatableInstanceCache getAnimatableInstanceCache()
+    {
+        return cache;
+    }
+}